Move command-line handling into CommandLine namespace
