+++ src/wix/WixToolset.Core.WindowsInstaller/CommandLine/InscribeSubcommand.cs
@@ -1,6 +1,6 @@
 // Copyright (c) .NET Foundation and contributors. All rights reserved. Licensed under the Microsoft Reciprocal License. See LICENSE.TXT file in the project root for full license information.
 
-namespace WixToolset.Core.WindowsInstaller
+namespace WixToolset.Core.WindowsInstaller.CommandLine
 {
     using System;
     using System.IO;
